refactor(jump_bar): extract Widgets
