+++ lib/widgets/jump_bar/jump_bar.dart
@@ -1,7 +1,7 @@
 import 'package:cabin_booking/model.dart';
-import 'package:cabin_booking/utils/dialog.dart';
 import 'package:cabin_booking/utils/string_extension.dart';
-import 'package:cabin_booking/widgets/jump_bar/booking_search_result.dart';
+import 'package:cabin_booking/widgets/jump_bar/jump_bar_field.dart';
+import 'package:cabin_booking/widgets/jump_bar/jump_bar_results.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:provider/provider.dart';
@@ -28,10 +28,10 @@ class _JumpBarState extends State<JumpBar> {
 
   double get _totalItems => _searchedBookings.length + _suggestedBookingsCount;
 
+  double get _maxHeight => _itemHeight * widget.maxVisibleItems;
+
   double get _height =>
       (_totalItems * _itemHeight).clamp(_itemHeight, _maxHeight);
-
-  double get _maxHeight => _itemHeight * widget.maxVisibleItems;
 
   @override
   void initState() {
@@ -90,11 +90,11 @@ class _JumpBarState extends State<JumpBar> {
                 child: Column(
                   mainAxisSize: MainAxisSize.min,
                   children: [
-                    _JumpBarField(controller: _controller),
+                    JumpBarField(controller: _controller),
                     if (_searchedBookings.isNotEmpty) const Divider(height: 0),
                     SizedBox(
                       height: _height,
-                      child: _JumpBarResults(
+                      child: JumpBarResults(
                         suggestedBooking: _suggestedBooking,
                         searchedBookings: _searchedBookings,
                         itemExtent: _itemHeight,
@@ -107,105 +107,6 @@ class _JumpBarState extends State<JumpBar> {
           ],
         ),
       ),
-    );
-  }
-}
-
-class _JumpBarResults extends StatelessWidget {
-  final SingleBooking? suggestedBooking;
-  final List<Booking>? searchedBookings;
-  final double itemExtent;
-
-  const _JumpBarResults({
-    super.key,
-    this.suggestedBooking,
-    this.searchedBookings,
-    required this.itemExtent,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return ListView(
-      itemExtent: itemExtent,
-      children: [
-        if (suggestedBooking != null)
-          _JumpBarItem(
-            icon: Icons.auto_awesome,
-            selected: true,
-            child: BookingSearchResult(booking: suggestedBooking!),
-            onTap: () async {
-              final cabinManager =
-                  Provider.of<CabinManager>(context, listen: false);
-
-              Navigator.of(context).pop();
-
-              return showNewBookingDialog(
-                context: context,
-                booking: suggestedBooking!.copyWith(
-                  cabinId:
-                      suggestedBooking!.cabinId ?? cabinManager.cabins.first.id,
-                ),
-                cabinManager: cabinManager,
-              );
-            },
-          ),
-        if (searchedBookings != null)
-          for (final booking in searchedBookings!)
-            _JumpBarItem(
-              icon: Icons.event,
-              child: BookingSearchResult(booking: booking),
-              onTap: () {},
-            ),
-      ],
-    );
-  }
-}
-
-class _JumpBarField extends StatelessWidget {
-  final TextEditingController? controller;
-
-  const _JumpBarField({super.key, this.controller});
-
-  @override
-  Widget build(BuildContext context) {
-    return TextFormField(
-      controller: controller,
-      autofocus: true,
-      decoration: const InputDecoration(
-        hintText: 'Type a new booking',
-        filled: false,
-        border: InputBorder.none,
-        icon: Padding(
-          padding: EdgeInsetsDirectional.only(start: 12),
-          child: Icon(Icons.search),
-        ),
-        contentPadding: EdgeInsetsDirectional.only(end: 16),
-      ),
-    );
-  }
-}
-
-class _JumpBarItem extends StatelessWidget {
-  final IconData? icon;
-  final Widget? child;
-  final bool selected;
-  final VoidCallback? onTap;
-
-  const _JumpBarItem({
-    super.key,
-    this.icon,
-    this.child,
-    this.selected = false,
-    this.onTap,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return ListTile(
-      leading: Icon(icon),
-      title: child,
-      selected: selected,
-      onTap: onTap,
     );
   }
 }
